+++ testData/editor/refactoring/HeredocTestCase/nestedHeredocRename/source_after.bash
@@ -2,6 +2,6 @@ if test; then
     if test; then
         echo && cat <<- EOF_RENAMED
     Heredoc content!
-EOF_RENAMED
+		EOF_RENAMED
     fi
 fi